pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "soja-produtividade"
version = "0.1.0"
description = "Pipeline reprodutível para previsão de produtividade de soja por município no Brasil"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Bruno"}
]
keywords = ["machine-learning", "agriculture", "soybean", "productivity", "brazil"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pandas>=2.0.0,<3.0.0",
    "numpy>=1.24.0,<2.0.0",
    "pyarrow>=14.0.0,<16.0.0",
    "requests>=2.31.0,<3.0.0",
    "geopandas>=0.14.0,<1.0.0",
    "shapely>=2.0.0,<3.0.0",
    "scikit-learn>=1.3.0,<2.0.0",
    "lightgbm>=4.0.0,<5.0.0",
    "shap>=0.44.0,<1.0.0",
    "matplotlib>=3.8.0,<4.0.0",
    "seaborn>=0.13.0,<1.0.0",
    "pyyaml>=6.0.0,<7.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "tqdm>=4.66.0,<5.0.0",
    "loguru>=0.7.0,<1.0.0",
    "pandera>=0.18.0,<1.0.0",
    "dvc>=3.30.0,<4.0.0",
    "scipy>=1.11.0,<2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0,<8.0.0",
    "pytest-cov>=4.1.0,<5.0.0",
    "ruff>=0.1.0,<1.0.0",
    "black>=23.0.0,<25.0.0",
    "pre-commit>=3.5.0,<4.0.0",
    "ipykernel>=6.25.0,<7.0.0",
    "jupyter>=1.0.0,<2.0.0",
]
dashboard = [
    "streamlit>=1.29.0,<2.0.0",
    "plotly>=5.18.0,<6.0.0",
]

[project.urls]
Repository = "https://github.com/bruno-portfolio/brazil-soybean-forecast"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
target-version = "py310"
line-length = 100
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without strict
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["ARG"]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
